Rewrite run-tests.sh to provide more flexibility
New version of run-tests.sh now takes command line arguments instead of
environment variables, always stores the results into a results
folder (in the top level of the tree, in parallel to the log files).

The new script can run GCC or GDB tests, GCC is run by default, but a
command line flag allows this to be changed.

Files Changed:

	* run-tests.sh: Rewrite to allow for running GCC or GDB tests,
	always stores the results away.  Takes command line arguments
	and not environment variables.