Split build-all.sh into build-tools.sh and build-targets.sh
This splits up the build-all.sh script to factor out the building
of the tools from the building of the verilator models and
any other targets. The clone scripts still clone all of the
repositories.

The new scripts offer some new behaviors. For build-targets.sh
the user is able to specify the directories containing the
sources for ri5cy and picvorv32, and for build-tools.sh the
user is able to specify the build directories for ri5cy
and picorv32 to be incorporated into the GDBServer.

Files changed:

	* README.md: Update to reflect splitting on build-all.sh into
	build-tools.sh and build-targets.sh. Also minor fixes.
	* build-all.sh: Remove script. Functionality reproduced by
	new scripts 'build-tools.sh' and 'build-targets.sh'.
	* build-isa-tests.sh: Fixes to formatting, addition of 'set -u'.
	Sources common.sh for functions common to other scripts.
	* build-targets.sh: New file which factors out the building
	of targets from the build-all.sh script.
	* build-testfloat.sh: Fixes to formatting, additon of 'set -u'.
	Sources common.sh for functions common to other scripts.
	* build-tools.sh: New file which factors out the building of
	the tool chain from the build-all.sh script.
	* common.sh: Factors out the common routines from the
	build-all.sh script.