Capture stdout from target when testing through GDB
When testing through GDB, targets can use File-I/O to use the
host to print to the terminal (as well as access files on the
host system). Some GCC tests rely on comparing the stdout from
the program, however the gdb dejagnu stub made no attempt to
capture this from the target.

This extends the GDB stub to capture the output from the
target and return it as the stdout from the running program.

There are some potential problems doing this:

1) The buffer used by expect is of limited size, so the
   output may become truncated if it exceeds the buffer
	 size.
2) Extra characters (eg newlines, the gdb prompt) not
   printed by the target may end up in the output, or
	 end up interleaved with the target's output.

Files changed:

	* dejagnu/gdb-stub-extra.exp: Capture stdout from GDB File-I/O
	when waiting for the target to halt.