Orthogonalise arguments to toolchain scripts
The arguments specifying ABI, target, and, arch were not
orthogonal, so it was possible to specify invalid combinations of
them (e.g. risc64-unknown-elf with rv32i arch). This commit
changes the options to:

  --with-xlen: 32 or 64
  --with-arch: Architecture and extensions only, e.g. i or ima.
  --with-abi: ilp, lp, or llp

These options are then translated to appropriate target, arch, and
abi strings for the toolchain configuration scripts.

Also included is a minor change for testing, because the gdbserver
is renamed to riscv<XLEN>-gdbserver.

Files Changed:

        * README.md: Update documentation of options.
        * build-tools.sh: Change --with-target to --with-xlen, and
        defaults for --with-arch and with-abi. Compute
        TARGET_TRIPLET, WITH_ARCH and WITH_ABI from options.
        * dejagnu/baseboards/riscv-gdbserver.exp: Call gdbserver
        as riscv<XLEN>-gdbserver.
        * run-tests.sh: Set env var RISCV_XLEN to help Dejagnu
        call gdbserver with correct xlen.