Rework the build-tools.sh script
The build-tools.sh script now has a --with-xlen option which can be 32
or 64.  Changing this will select suitable defaults for the
--with-target and --with-arch flags.

The --with-arch flag now takes only the architectural features to
include, the 'rv32' or 'rv64' prefix is added automatically based on the
value passed in --with-xlen.

The --with-abi flag can now be dropped in most cases, we instead rely on
GCC to pick a suitable default.  The flag does still exists if the user
wants to force a particular (non-standard) choice.

The --with-target flag will also take a suitable default based on the
value passed in --with-xlen, so in most cases this will not be needed
either.

We also build and install the C++ libraries now as part of the GCC stage
2 build.

Files Changed:

	* build-tools.sh: Rework to handle new options and default
          selection.
	* common.sh: Add a couple of new helper functions.