Don't overwrite build log if building in parallel
If multiple builds run in parallel then don't overwrite the build log
file.  The build log file is currently created using a timestamp to the
nearest minute, if two builds start within one minute then the
timestamps will match and the build log will be overridden.

Simply adding the second to the filename is not a good solution, if
the builds are started by a script then they may all start within the
same second.

The solution is to add a unique key to the end of the filename using
mktemp.

Files Changed:

	* build-tools.sh: Add unique key to the build log name.