Overhaul how riscv baseboard are managed
This commit does two things, first it moves the riscv-sim.exp and
riscv-spike.exp baseboards out of the dejagnu repository, and places
them in the toolchain repository.  This is a better strategy in the
long term as, it means we no longer need to maintain the separate
dejagnu repository at all.

This commit also does a massive rewrite of how the gdbserver baseboard
operates.  Most of the custom code is now gone, and instead, we load
the standard gdbserver library, and override a couple of functions to
get our gdbserver working withing GDB's testsuite, so far, the new
strategty appears to be working well.

Files Changed:

	* dejagnu/baseboards/riscv-gdbserver.exp: Rewrite to use new
	gdbserver strategy.
	* dejagnu/baseboards/riscv-sim.exp: New file.  Copied from old
	dejagnu repository.
	* dejagnu/baseboards/riscv-spike.exp: New file.  Copied from old
	dejagnu repository.
	* dejagnu/gdb-extra.exp: Deleted.
	* dejagnu/gdb-stub-extra.exp: Deleted.
	* dejagnu/remote-extra.exp: Deleted.
	* dejagnu/riscv-gdbserver-support.exp: New file.