feedom-e310: Reorder the linker flags
In order to get C++ tests building, reorder how libraries are pulled
into the build.  Prevent the GCC wrapper adding libgcc and libc to the
end of the link line (before libwrap) using -nostdlib, then explicitly
pull those libraries in ourselves in the correct order relative to
libwrap (that is before).

I've thrown everything into start-group/end-group for now, which might
be overkill, we could investigate this more later if we want.