Add options to benchmark-beebs.py
Options to configure the architectures and configurations are
benchmarked are added, used as per the help text:

"""
$ ./benchmark-beebs.py --help
usage: benchmark-beebs.py [-h] [--arches ARCH [ARCH ...]]
                          [--configs CONFIG [CONFIG ...]]

Run BEEBS Benchmarks for a given set of configurations on a given set of
architectures. The default is to run all configurations on all architectures.

Available configurations are:

    baseline nocrt nolibc nolibc-nolibgcc nolibc-nolibgcc-nolibm

Available architectures are:

    riscv arm arc

optional arguments:
  -h, --help            show this help message and exit
  --arches ARCH [ARCH ...]
                        Target architectures
  --configs CONFIG [CONFIG ...]
                        Configurations
"""

ChangeLog:

        * benchmark-beebs.py: Add argument parsing.